+++ include/auton.h
@@ -1,0 +1,8 @@
+// auton.h
+
+#ifndef AUTON_H
+#define AUTON_H
+
+void autonomous(void);
+
+#endif // AUTONOMOUS_H
