+++ include/usercontrol.h
@@ -1,0 +1,8 @@
+// usercontrol.h
+
+#ifndef USERCONTROL_H
+#define USERCONTROL_H
+
+void usercontrol(void);
+
+#endif // USERCONTROL_H
